Moving autocomplete_selector out of the widgets library
