+++ components/interactive_dialog/dialog_element.jsx
@@ -11,7 +11,7 @@ import GenericUserProvider from 'components/suggestion/generic_user_provider.jsx
 import GenericChannelProvider from 'components/suggestion/generic_channel_provider.jsx';
 
 import TextSetting from 'components/widgets/settings/text_setting';
-import AutocompleteSelector from 'components/widgets/settings/autocomplete_selector';
+import AutocompleteSelector from 'components/autocomplete_selector';
 import BoolSetting from 'components/widgets/settings/bool_setting';
 
 const TEXT_DEFAULT_MAX_LENGTH = 150;
